Refactoring: Renamed ecu_config_reader to ecu_config
