+++ can_logger.py
@@ -1,5 +1,5 @@
 import can
-import ecu_config_reader
+import ecu_config
 import os
 import datetime
 
@@ -7,7 +7,7 @@ LOG_FILE_NAME_FORMAT = 'can_%y%m%d%H%M%S.log'
 
 MAX_LOG_FILE_SIZE = 1500000  # bytes
 
-CAN_INTERFACE = ecu_config_reader.get_can_interface()
+CAN_INTERFACE = ecu_config.get_can_interface()
 
 BUS_TYPE = "socketcan_native"
 
@@ -32,19 +32,19 @@ def get_filters():
 
 
 def get_can_ids():
-    ecu_can_ids = ecu_config_reader.get_ecu_addresses()
+    ecu_can_ids = ecu_config.get_ecu_addresses()
     tester_can_ids = []
     for ecu_can_id in ecu_can_ids:
         tester_can_ids.append(ecu_can_id+8)
     return ecu_can_ids + tester_can_ids
 
 
+def create_new_file_path_if_size_exceeded(file_path):
+    if os.path.exists(file_path):
+        if os.path.getsize(file_path) > MAX_LOG_FILE_SIZE:
+            file_path = create_file_path()
+    return file_path
+
+
 def create_file_path():
     return os.path.join(os.path.dirname(__file__), datetime.datetime.now().strftime(LOG_FILE_NAME_FORMAT))
-
-
-def create_new_file_path_if_size_exceeded(log_file):
-    if os.path.exists(log_file):
-        if os.path.getsize(log_file) > MAX_LOG_FILE_SIZE:
-            log_file = create_file_path()
-    return log_file
